requirements.txt
pyephem
pytz
pandas
numpy
jinja2==3.1.3
streamlit
notebook
jupyterhub
pandas
matplotlib
streamlit_folium